Updates 17.03.18
New classes
- Map
	- Creates and stores a 2D array of Rooms
- Room
	- Stores information about a room
		- All obstacles, enemies, buttons etc.
- RandomNumberHelper
	- Creates random numbers, but every number has its own probablity
- ChargeBot
	- An enemy which charges the player if the player is within the bot's sector
- ChargeBotCtrl
	- A Controller for the ChargeBot
- StatsPanel
	- A panel which is left of the main frame
	- Displays the stats of the player's ship.
- InfoPanel
	- A superclass which StatsPanel and EastPanel inherit from.

Classes Changed
- Doors
	- Now closes upon entering a new room
Other
- Made a randmo map generator
	- Generates a random map every time a new game is launched

Error fixed:
- None

+++ src/Assignment/GameObjects/ObjectStats.java
@@ -1,6 +1,7 @@
 package Assignment.GameObjects;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -40,7 +41,6 @@ public class ObjectStats {
     ObjectStats(Map<String, Integer> initMap){
         stats = new HashMap<>(initMap);
     }
-
 
     void newLife(){
         stats.put(ARMOUR, getMaxArmour());
@@ -84,7 +84,7 @@ public class ObjectStats {
     public int getLivesRemaining() {
         return stats.get(LIVES_REMAINING);
     }
-    public long getFireRate() {
+    public int getFireRate() {
         return  stats.get(FIRE_RATE);
     }
